+++ attributes/default.rb
@@ -7,3 +7,6 @@ default['topo']['topologies_data_bag'] = 'topologies'
 # topo.node_type is used to look for node data if no data is
 # found in the topology data bag item under node.name
 default['topo']['node_type'] = 'default'
+
+# when to delete the validation key ('setup', 'shutdown', 'never')
+default['topo']['delete_validation_key'] = 'shutdown'