+++ example/testapp/recipes/appserver.rb
@@ -1,0 +1,8 @@
+#
+# Cookbook Name:: testapp
+# Recipe:: appserver
+#
+# Copyright 2015, ThirdWave Insights
+#
+
+include_recipe "nodejs::default"