+++ metadata.rb
@@ -4,4 +4,5 @@ maintainer_email 'christine_draper@thirdwaveinsights.com'
 license 'Apache 2.0'
 description 'Sets node configuration from a topology data bag'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
-version '0.1.2'
+version '0.1.3'
+supports 'ubuntu'