+++ spec/spec_helper.rb
@@ -9,3 +9,17 @@ def create_chef_node(resource_name)
     :create,
     resource_name)
 end
+
+def delete_chef_node(resource_name)
+  ChefSpec::Matchers::ResourceMatcher.new(
+    :chef_node,
+    :delete,
+    resource_name)
+end
+
+def delete_chef_client(resource_name)
+  ChefSpec::Matchers::ResourceMatcher.new(
+    :chef_client,
+    :delete,
+    resource_name)
+end