+++ test/integration/blueprint/serverspec/blueprint_spec.rb
@@ -5,6 +5,7 @@ describe 'topo::setup_node' do
   # http://serverspec.org/resource_types.html
   context 'using blueprint' do
     it 'converges using test topology' do
+      skip 'Should have  node name dbserver01'
       skip 'Should have set runlist to include blueprint-test cookbook'
     end
   end